Improve build time + misc other refactorings

+++ src/iso/table_of_contents.h
@@ -49,7 +49,7 @@ struct LevelInfo {
 	Opt<LevelWadInfo> audio;
 	Opt<LevelWadInfo> scene;
 	s32 level_table_index; // Not used for writing.
-	s64 level_table_entry_offset = 0; // Not used for writing.
+	s32 level_table_entry_offset = 0; // Not used for writing.
 };
 
 struct table_of_contents {
